feat: catalog style toggle (classic/new) with admin mode
Add configurable catalog layout via ui-config.json:
- "catalog.style": "classic" — original Nitro layout (tabs, grid)
- "catalog.style": "new" — modern layout (sidebar icons, breadcrumb,
  favorites, search bar)

Both layouts include full admin mode:
- Edit/create/delete pages and offers
- Toggle visibility and enabled state
- Reorder items
- Publish catalog (reloads for all connected clients)
- Admin mode toggle visible only for moderators

CatalogView.tsx acts as a router that reads the config and loads
either CatalogClassicView or CatalogModernView.

Also fixes vite.config.mjs alias for @nitrots/nitro-renderer to
resolve through renderer3 source instead of stale node_modules copy.

Requires: renderer3 CatalogAdmin packets + emulator handlers
