+++ public/ui-config.json
@@ -1348,8 +1348,9 @@
     "catalog.asset.url": "${image.library.url}catalogue",
     "catalog.asset.image.url": "${catalog.asset.url}/%name%.gif",
     "catalog.asset.icon.url": "${catalog.asset.url}/icon_%name%.png",
-    "catalog.tab.icons": false,
+    "catalog.tab.icons": true,
     "catalog.headers": false,
+    "catalog.style": "old",
     "chat.input.maxlength": 100,
     "chat.styles.disabled": [],
     "chat.styles": [{
